improved extendAtomicData, added getHierView support to AtomMap
